Ipv6Test: rename to `CheckIpv6Test`
... as the remaining tests relate directly to the `check_ipv6()` method.

Including re-ordering the methods in the class to follow the order of the code in the method under test.
